refactor(runtime): split scoped_session_auth.rs god object into modules
scoped_session_auth.rs (744 LOC) -> scoped_session_auth/: mod.rs (55) keeps
the module doc + shared imports + re-exports; config.rs (68) the public DTOs
(ScopedAuthMode/Config/Resource/SessionClaims/…); state.rs (159) the challenge/
session entries + time+load-gated GC; service.rs (236) the ScopedAuthService
impl + token issuance. Public API preserved via pub use so
sandbox_runtime::scoped_session_auth::{ScopedAuthService,ScopedAuthConfig,…}
still resolves for tests/bench_regression.rs and the bench. Pure code motion,
no logic/crypto change; max module 236 LOC.
